requirements.txt
colour-science
wgpu-py
numpy
pillow
scipy
six
